pyproject.toml
[project]
name = "rapids-meta-orchestrator"
version = "0.1.0"
description = "RAPIDS Meta-Orchestrator: Multi-Project Workspace System with Claude Agent SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "claude-agent-sdk>=0.1.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["orchestrator/backend/tests", "tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
